Adding some modifications to use more generic rest with patch and find

+++ src/Controller/DeleteResourceController.php
@@ -2,22 +2,24 @@
 
 namespace JDesrosiers\Resourceful\Controller;
 
-use Doctrine\Common\Cache\Cache;
+use Silex\Application;
 use Symfony\Component\HttpFoundation\Request;
 use Symfony\Component\HttpFoundation\Response;
+use Symfony\Component\HttpKernel\Exception\NotFoundHttpException;
 use Symfony\Component\HttpKernel\Exception\ServiceUnavailableHttpException;
 
-class DeleteResourceController
+/**
+ * Class DeleteResourceController
+ * @package JDesrosiers\Resourceful\Controller
+ */
+class DeleteResourceController extends AbstractResourceController
 {
-    private $service;
-
-    public function __construct(Cache $service)
+    public function __invoke(Application $app, Request $request, $id)
     {
-        $this->service = $service;
-    }
+        if (!$this->service->contains($request->getRequestUri())) {
+            throw new NotFoundHttpException("Not Found");
+        }
 
-    public function __invoke(Request $request)
-    {
         if ($this->service->delete($request->getRequestURI()) === false) {
             throw new ServiceUnavailableHttpException(null, "Failed to delete resource");
         }
